Fix(chat): Emoji picker appearing out of bounds on mobile
